refactor : 컨포넌트 어노테이션을 빼고 WebMvcConfiguration.java 에서 빈으로 정의해서 명시적으로 인프라에 있는 리졸버를 로딩
