+++ src/main/java/nextstep/infrastructure/AuthenticationPrincipal.java
@@ -1,4 +1,4 @@
-package nextstep.config;
+package nextstep.infrastructure;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
