--- src/main/java/nextstep/infrastructure/LoginMember.java
@@ -1,17 +1,0 @@
-package nextstep.infrastructure;
-
-public class LoginMember {
-    private Long id;
-
-    private LoginMember(Long id) {
-        this.id = id;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public static LoginMember from(String id) {
-        return new LoginMember(Long.parseLong(id));
-    }
-}
