+++ src/main/java/nextstep/infrastructure/UserContextHolder.java
@@ -1,0 +1,17 @@
+package nextstep.infrastructure;
+
+public class UserContextHolder {
+    private Long id;
+
+    private UserContextHolder(Long id) {
+        this.id = id;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public static UserContextHolder from(String id) {
+        return new UserContextHolder(Long.parseLong(id));
+    }
+}
