+++ src/main/java/nextstep/interfaces/ExceptionHandlers.java
@@ -1,5 +1,7 @@
-package nextstep.support;
+package nextstep.interfaces;
 
+import nextstep.support.DuplicateEntityException;
+import nextstep.support.NotExistEntityException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
